erp5_trade: ignore possible KeyError with catalog results.
That can happen three Zope transactions happen like the following order :

* Zope1: create a Sale Packing List
  * Zope2: build activity started
* Zope1: transaction committed
    * Zope3: index activity started
    * Zope3: MariaDB transaction committed
  * Zope2: portal_catalog() finds a row

Here, Zope2 cannot find the Sale Packing List, because its transaction started before the commit of Zope1, thus brain.getObject() raises an KeyError.

It can happen elsewhere and it is normally not a big issue, as the activity will be retried.

But for benchmark (see software/erp5/test/test/benchmarks.py in SlapOS repository), the performance differs much if such KeyError happens a lot or not.
This commit is to ignore possible KeyError in benchmark so that we can get stable results.