Inventory API: Do not use regex for measure
`concat(stock.variation_text,'\n') REGEXP measure.variation` is slow.
This introduces an alternative approach of indexing the first 4
variations directly in columns and use a traditional join instead or
a regexp match.
In a real production case of a 2 seconds query, this was one order of
magnitude faster.

This has some limitations:
  - it does not support more than 4 variations
  - it does not support optional variations

The previous regexp based implementation is kept in this patch, so that
it's easy to rollback this change if side effects are found.