core,hal_json_style: avoid using HTTPRequest.keys()
ZPublisher.HTTPRequest.HTTPRequest.keys() sorts the keys, which is a
TypeError on python3 when request have non-string keys.

It's not clear if non-string are really legal, other things such as
request.text() also error when it encounters some request keys that are
not strings, because _filterPasswordFields expects them to be str.

Some custom dialog code sometimes use request.other as a cache storage
for the duration of the request, so it's safer if we can just tolerate
it - especially because the pattern was iterating on request keys to
remove them from request.form, we can directly iterate on request.form
keys.