ERP5Type: only check consistency up to next decision workflow
When an object has sub-objects, consistency checking on the object was always
checking consistency of all sub-objects. This does not makes sense for
sub-objects having an independent decision (ie. not interaction, not edit)
workflow.

For instance, one should be able to validate a Person even when its
Assignments, which have their own Assignment Workflow, are inconsistent,
because in any case the check can be bypassed by validating the Person first,
then adding an Assignment, which is not going to invalidate the Person. Also,
validating the Person has no actual impact on security before the Assignment is
actually started, operation which will check consistency of the Assignment.

On the other hand, we still want to propagate consistency of sub-objects in
cases where they do not have an independent decision workflow. For instance,
consistency of an Order should always ensure consistency of all the lines,
since lines have no other workflow than edit or interaction workflow.

This commit therefore introduces this logic in ERP5: for all sub-objects, first
check if there is a decision workflow before extending the parent object
consistency with consistency of the sub-object.