XMLExportImport: fix persistent_id serialized as <string> on py2
On Python 2, persistent IDs are serialized using the BYTE opcode in ZODB
records, but only after ZODB.serialize.ObjectWriter saved them in a
storage, because storage's new_oid method always returns a str on py2
and it's the ObjectWriter.persistent_id method [1] which takes care of
converting the persistent_id from str to zodbpickle.binary.

When BusinessTemplate exports an object to XML, the following logic
happens:
 1. use _getCopy to make an copy of the object and all the persistent
  objects referred by this object, recursively
 2. call removeProperties on the copy
 3. commit a transaction savepoint
 4. use exportXML on the copy

In step 1. the copy and all the persistent references still have str
persistent_id. After step 3. the objects that were modified in step 2
are saved in the transaction's savepoint storage ( ZODB.Connection.TmpStore )
and have zodbpickle.binary persistent_id for all their referenced
objects, but if the referenced objects themselves also reference other
objects and were not modified before the savepoint, they still have
their persistent_id as str.

Because of this, when an object o1 is exported, if it holds references
to other persistent object o2, o2's persistent_id is BYTES, as expected,
but if o2 has reference to another o3 persistent object, o3's
persistent_id is still str and it is serialized in XML as something like
that:

    <persistent> <string encoding="base64">AAAAAAAAAAI=</string> </persistent>

To prevent this, we explictly convert the persistent_id to
zodbpickle.binary in XMLExportImport.

[1]: https://github.com/zopefoundation/ZODB/blob/5.8.1/src/ZODB/serialize.py#L383