BusinessTemplate: special case Page Template's title serialization on py2
On py2, ZopePageTemplate's title are unicode objects, that is serialized
in business template XML as <unicode>, but on py3 business template it's
serialized as <string>, because we no longer have <unicode> on py3.

To achieve the same business template format, adjust the page templates
export on py2 so that they appear as <string> in the XML and do the
reverse operation to decode the title to unicode when installing the
business template.