XMLExportImport: always output a <state> for objects with empty __dict__
Without this special handling, on py3 object without state would be
exported like this:

    <pickle>
      <object>
        <klass>
          <global name="DummyEmptyClass" module="testXMLPickle"/>
        </klass>
        <tuple/>
      </object>
    </pickle>

which is different from the expected py2 output:

    <pickle>
      <object>
        <klass>
          <global name="DummyEmptyClass" module="testXMLPickle"/>
        </klass>
        <tuple/>
        <state>
          <dictionary/>
        </state>
      </object>
    </pickle>

because the state is missing. This is because object.__reduce_ex__
returns a state of None on py3 and this does not appear in the pickle
stream, an empty object on py3 is only:

    0: \x80 PROTO      3
    2: c    GLOBAL     'testXMLPickle DummyEmptyClass'
   33: q    BINPUT     0
   35: )    EMPTY_TUPLE
   36: \x81 NEWOBJ
   37: q    BINPUT     1
   39: .    STOP

on py2, it is:

    0: \x80 PROTO      3
    2: c    GLOBAL     'testXMLPickle DummyEmptyClass'
   33: q    BINPUT     0
   35: )    EMPTY_TUPLE
   36: \x81 NEWOBJ
   37: q    BINPUT     1
   39: }    EMPTY_DICT
   40: q    BINPUT     2
   42: b    BUILD
   43: .    STOP

Also add a few missing unit tests.