payroll_l10n_fr: Rename DSNChangeBlock.data into previous_value
On py3, we cannot use properties of type bytes using a string field in
the user interface, bytes are only for stream of data that cannot always
be represented as text.

To make this module fully compatible with py3, we had to change the
property to be of type string. To prevent confusion with existing data
property, that is bytes, we use a new property. Because this module is
not used much, we decided not to implement any data migration.